[REFACTOR] 파일, 폴더, 컴포넌트 이름 변경
- 파일, 폴더, 컴포넌트 이름 변경
   컨벤션 변경에 따른 이름 변경
