+++ frontEnd/src/components/Router/Router.js
@@ -7,7 +7,7 @@ import {
   useHistory,
 } from 'react-router-dom';
 
-import IssueListPage from '../../pages/issueListPage/issueListPage';
+import IssueListPage from '../../pages/issue-list-page/IssueListPage';
 import Login from '../../pages/login';
 
 const IssueTrackerRouter = ({token}) => {
