+++ frontEnd/src/components/checkbox-with-description/CheckBoxWithPresenter.js
