+++ frontEnd/src/components/checkbox-with-description/index.js
