+++ frontEnd/src/components/label-btn/LabelBtnContainer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LabelBtnPresenter from './labelBtnPresenter';
+import LabelBtnPresenter from './LabelBtnPresenter';
 const LabelBtnContainer = () => {
   return <LabelBtnPresenter labelClick={() => {}} count={8} />;
 };
