+++ frontEnd/src/components/label-btn/index.js
@@ -1,0 +1,3 @@
+import LabelBtnContainer from './LabelBtnContainer';
+
+export default LabelBtnContainer;
