--- frontEnd/src/components/labelBtn/index.js
@@ -1,3 +1,0 @@
-import LabelBtnContainer from './labelBtnContainer';
-
-export default LabelBtnContainer;
