+++ frontEnd/src/components/milestone-btn/index.js
@@ -1,0 +1,3 @@
+import MilestoneBtnContainer from './MilestoneBtnContainer';
+
+export default MilestoneBtnContainer;
