--- frontEnd/src/components/milestoneBtn/index.js
@@ -1,3 +1,0 @@
-import MilestoneBtnContainer from './milestoneBtnContainer';
-
-export default MilestoneBtnContainer;
