+++ frontEnd/src/components/new-Issue-btn/index.js
@@ -1,0 +1,3 @@
+import NewIssueBtnContainer from './NewIssueBtnContainer';
+
+export default NewIssueBtnContainer;
