--- frontEnd/src/components/newIssueBtn/index.js
@@ -1,3 +1,0 @@
-import NewIssueBtnContainer from './newIssueBtnContainer';
-
-export default NewIssueBtnContainer;
