+++ frontEnd/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDom from 'react-dom';
 
-import App from './components/App/App.js';
+import App from './components/app/App.js';
 
 ReactDom.render(<App />, document.getElementById('app'));
