+++ frontEnd/src/pages/issue-list-page/issue-main/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import IssueList from './IssueList';
+import IssueList from './issue-list';
 import styled from 'styled-components';
-import IssueMainHeader from './IssueMainHeader';
+import IssueMainHeader from './issue-main-header';
 // 이곳에 ContextAPI 나중에 추가할것!
 
 const issueMainWidth = '90%';
